requirements.txt
jupyter                   1.0.0           py311h267d04e_8    conda-forge
jupyter_client            8.2.0              pyhd8ed1ab_0    conda-forge
jupyter_console           6.6.3              pyhd8ed1ab_0    conda-forge
jupyter_core              5.3.0           py311h267d04e_0    conda-forge
matplotlib                3.7.1           py311ha1ab1f8_0    conda-forge
matplotlib-base           3.7.1           py311h99a5f44_0    conda-forge
numpy                     1.24.2          py311h60f8152_0    conda-forge
owlready2                 0.44                     pypi_0    pypi
pandas                    2.0.0           py311h4eec4a9_0    conda-forge
pip                       23.1.2             pyhd8ed1ab_0    conda-forge
pytest                    8.0.2                    pypi_0    pypi
python                    3.11.3          h1456518_0_cpython    conda-forge
python-dateutil           2.8.2              pyhd8ed1ab_0    conda-forge
python-fastjsonschema     2.16.3             pyhd8ed1ab_0    conda-forge
python-json-logger        2.0.7              pyhd8ed1ab_0    conda-forge
python-tzdata             2023.3             pyhd8ed1ab_0    conda-forge
python-xxhash             3.2.0           py311he2be06e_0    conda-forge
python_abi                3.11                    3_cp311    conda-forge
pytorch                   2.0.0           cpu_py311hb4bb8ad_0    conda-forge
scikit-learn              1.2.2                    pypi_0    pypi
scipy                     1.10.1                   pypi_0    pypi
torchtext                 0.6.0                      py_1    pytorch
tornado                   6.3             py311he2be06e_0    conda-forge
tqdm                      4.66.2             pyhd8ed1ab_0    conda-forge